Require IRQ number to bhyve VirtIO emulation
The IRQ of each device is now mandatory. If a device does not use IRQs,
the specified number should be -1

+++ usr.sbin/bhyve/mmio/devemu.h
@@ -51,7 +51,7 @@ struct devemu_inst {
 
 	struct {
 		enum lintr_stat	state;
-		uint32_t	irq;
+		int64_t		irq;
 		pthread_mutex_t	lock;
 	} di_lintr;
